requirements.txt
langchain==0.0.352
transformers==4.35.2
sentence-transformers==2.2.2
chromadb==0.4.18
faiss-cpu==1.7.4
pypdf==3.17.1
torch==2.7.1
openai==1.3.7
streamlit==1.29.0
python-dotenv==1.0.0
pyyaml==6.0.1
huggingface_hub==0.19.4
